refactor: use misc common utilities
